Rename package to new name
