--- src/main/java/nl/mawoo/migratejs/exceptions/CantFindLibrary.java
@@ -1,16 +1,0 @@
-package nl.mawoo.migratejs.exceptions;
-
-/**
- * If a user want to use the use() function.
- *
- * @author Bob van der Valk
- */
-public class CantFindLibrary extends WCMScriptException {
-    public CantFindLibrary(String message) {
-        super(message);
-    }
-
-    public CantFindLibrary(String message, Throwable cause) {
-        super(message, cause);
-    }
-}
