--- src/main/java/nl/mawoo/migratejs/exceptions/CantSaveFile.java
@@ -1,14 +1,0 @@
-package nl.mawoo.migratejs.exceptions;
-
-/**
- * Created by Bob on 14-3-2016.
- */
-public class CantSaveFile extends WCMScriptException {
-    public CantSaveFile(String message) {
-        super(message);
-    }
-
-    public CantSaveFile(String message, Throwable cause) {
-        super(message, cause);
-    }
-}
