--- src/main/java/nl/mawoo/migratejs/exceptions/CantUseFile.java
@@ -1,14 +1,0 @@
-package nl.mawoo.migratejs.exceptions;
-
-/**
- * Created by Bob on 6-3-2016.
- */
-public class CantUseFile extends WCMScriptException {
-    public CantUseFile(String message) {
-        super(message);
-    }
-
-    public CantUseFile(String message, Throwable cause) {
-        super(message, cause);
-    }
-}
