--- src/main/java/nl/mawoo/migratejs/exceptions/CustomScript.java
@@ -1,14 +1,0 @@
-package nl.mawoo.migratejs.exceptions;
-
-/**
- * Created by bobva on 09-04-16.
- */
-public class CustomScript extends WCMScriptException {
-    public CustomScript(String message) {
-        super(message);
-    }
-
-    public CustomScript(String message, Throwable cause) {
-        super(message, cause);
-    }
-}
