--- src/main/java/nl/mawoo/migratejs/exceptions/WCMScriptException.java
@@ -1,14 +1,0 @@
-package nl.mawoo.migratejs.exceptions;
-
-/**
- * Created by Bob on 3-3-2016.
- */
-public class WCMScriptException extends RuntimeException {
-    public WCMScriptException(String message) {
-        super(message);
-    }
-
-    public WCMScriptException(String message, Throwable cause) {
-        super(message, cause);
-    }
-}
