+++ src/main/java/nl/mawoo/wcmscript/exceptions/CantSaveFile.java
@@ -1,0 +1,14 @@
+package nl.mawoo.wcmscript.exceptions;
+
+/**
+ * Created by Bob on 14-3-2016.
+ */
+public class CantSaveFile extends WCMScriptException {
+    public CantSaveFile(String message) {
+        super(message);
+    }
+
+    public CantSaveFile(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
