+++ src/main/java/nl/mawoo/wcmscript/exceptions/MongoDB.java
@@ -1,0 +1,15 @@
+package nl.mawoo.wcmscript.exceptions;
+
+/**
+ * Created by Bob on 15-3-2016.
+ */
+public class MongoDB extends WCMScriptException {
+
+    public MongoDB(String message) {
+        super(message);
+    }
+
+    public MongoDB(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
