+++ src/main/java/nl/mawoo/wcmscript/exceptions/SQLError.java
@@ -1,0 +1,14 @@
+package nl.mawoo.wcmscript.exceptions;
+
+/**
+ * Created by Bob on 15-3-2016.
+ */
+public class SQLError extends WCMScriptException {
+    public SQLError(String message) {
+        super(message);
+    }
+
+    public SQLError(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
