+++ src/main/java/nl/mawoo/wcmscript/exceptions/WCMScriptException.java
@@ -1,0 +1,14 @@
+package nl.mawoo.wcmscript.exceptions;
+
+/**
+ * Created by Bob on 3-3-2016.
+ */
+public class WCMScriptException extends RuntimeException {
+    public WCMScriptException(String message) {
+        super(message);
+    }
+
+    public WCMScriptException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}
