+++ src/main/resources/dbconnector.js
@@ -7,7 +7,7 @@
  */
 var DbConnector = function(connString, username, password) {
     try {
-        var DBConnector = Java.type("nl.mawoo.migratejs.extend.dbconnector.DbConnector");
+        var DBConnector = Java.type("nl.mawoo.wcmscript.extend.dbconnector.DbConnector");
         this.connector = new DBConnector(connString, username, password);
     } catch(e) {
         print("Could not connect to the database \n Check your connection settings.");
