+++ src/main/resources/mongodb.js
@@ -3,7 +3,7 @@
  * @constructor
  */
 var MongoConnector = function(database) {
-    var mongoClass = Java.type("nl.mawoo.migratejs.extend.dbconnector.MongoConnector");
+    var mongoClass = Java.type("nl.mawoo.wcmscript.extend.dbconnector.MongoConnector");
     return new mongoClass(database);
 };
